% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.r
\name{build_edges}
\alias{build_edges}
\title{Build edge data frame between documents}
\usage{
build_edges(candidates, shingles, threshold = 0.8)
}
\arguments{
\item{candidates}{list of buckets with document ids from \code{lsh}}

\item{shingles}{list of documents and their shingles from \code{shingle}}

\item{threshold}{jaccard similarity threshold}
}
\description{
Builds an edge list based on whether documents are similar based on their
shingles. Uses the candidates list to make comparing document much faster.
}
