% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measures.R
\name{cosine}
\alias{cosine}
\title{Fast cosine similarity calculation.}
\usage{
cosine(m)
}
\arguments{
\item{m}{- sparse \code{\link{dgCMatrix}} or dense input matrix.
\code{cosine} calculates all pair similarities between rows of this matrix.}
}
\description{
Fast cosine similarity calculation.
}
\examples{
i <- c(1,3:8, 1:6); j <- c(2,9,6:10,6:1); x <- c(6 * (1:6), 7 * (1:7))
m <- sparseMatrix(i, j, x = x)
cosine(m)
}

