% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measures.R
\name{cosine_signatures}
\alias{cosine_signatures}
\title{similarity between two vectors.}
\usage{
cosine_signatures(x, y)
}
\arguments{
\item{x}{- \link{vector}}

\item{y}{- \link{vector}}
}
\value{
\link{numeric} cosine similarity coefficient
}
\description{
similarity between two vectors.
}

