% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix.R
\name{create_dtm}
\alias{create_dtm}
\title{Construct binary Document-Term matrix
from input list of elements.}
\usage{
create_dtm(lst, format = c("lil", "ngC"))
}
\arguments{
\item{lst}{input list of values in a form of
\code{\link{list}} of \code{\link{vector}}s}

\item{format}{desired format of result.
\code{lil} - List of Lists format
\code{ngC} - \code{\link{ngCMatrix-class}}}
}
\description{
This function is specially useful when creating
Document-Term matrix from input list of sets for "minhashing"
}
\examples{
generate artificial elements of sets
elems <- sapply(1:100, function(z)
paste(sample(letters, sample(3:7), replace = T), collapse=''))
generate artificial sets
sets <-  lapply(1:1e3, function(z) sample(elems, sample(20:100)))
dtm_lil <- create_dtm(sets, format = "lil")
}

