% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix.R
\name{create_tdm}
\alias{create_tdm}
\title{Construct Term-Document matrix (in "List of Lists" form) from input list of elements.}
\usage{
create_tdm(lst)
}
\arguments{
\item{lst}{input list of values in a form of \code{\link{list}} of \code{\link{vector}}s}
}
\description{
Construct Term-Document matrix (in "List of Lists" form) from input list of elements.
}

