% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsh.R
\name{get_candidate_pairs}
\alias{get_candidate_pairs}
\title{Calculating candidate pairs using locality sensitive hashing.}
\usage{
get_candidate_pairs(signature_matrix, bands_number, similarity,
  verbose = TRUE)
}
\arguments{
\item{signature_matrix}{input signature matrix - \code{\link{integer}} \code{\link{matrix}}}

\item{bands_number}{number of bands for LSH algorithm.}

\item{similarity}{target value of jacard similarity we are looking for.}

\item{verbose}{- \link{logical} print useful algorithm information.}
}
\value{
pairs of candidates with similarity => \code{similarity} -
\code{\link{data.table}} with 3 colums: index1, index2, N -
index of first candidate, index of second candidate,
and number of buckets where they share same value. The latter provided
only for information.
(Intutition is following: the bigger N - the bigger similarity)
}
\description{
Calculating candidate pairs using locality sensitive hashing.
}
\examples{
sets <- lapply(1:10, function(x) sample(letters, sample(5:15)))
# add set similar to first set to the end of list
sets <- c(sets, list(c(sets[[1]], sample(letters, 5))))
sm <- get_signature_matrix(sets, 12, cores = 4)
get_candidate_pairs(sm, 6, 0.9)
}

