% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signatures.R
\name{get_hash_matrix}
\alias{get_hash_matrix}
\title{Calculating hashing matrix for given similarity measure.}
\usage{
get_hash_matrix(dtm, hashfun_number, measure, cores)
}
\arguments{
\item{hashfun_number}{- number of projections to use for approximation. Involves
tradeoff: larger is better, but computattionally more expensive and requires more RAM}

\item{measure}{Similarity measure}

\item{cores}{number of CPU cores to use for hashing. (related only for "jaccard" measure)}

\item{terms_size}{number of}
}
\description{
For "jaccard" we generate construct hashing matrix
which simulate to permutation of rows. For "cosine" we generate random
hyperplanes.
}

