% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_curve.R
\name{get_s_curve}
\alias{get_s_curve}
\title{Calculates/plots "S-curve"}
\usage{
get_s_curve(number_hashfun, n_bands_min = 1, n_rows_per_band_min = 1,
  plot = interactive())
}
\arguments{
\item{number_hashfun}{guess about number of hash functions to use}

\item{n_bands_min}{don't plot s-curves for number of bands less than \code{n_bands_min}}

\item{n_rows_per_band_min}{don't plot s-curves for number of rows less than \code{n_rows_per_band_min}}

\item{plot}{logical, whether to plot s-curves.}
}
\description{
Calculates "S-curve". This is used to pick number of hash functions and bands which
provides tradeoff between precision and recall for approximate near neighbor search.
}
