% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.r
\name{group_edges}
\alias{group_edges}
\title{Group documents together into a cluster}
\usage{
group_edges(edges, docs = NULL)
}
\arguments{
\item{edges}{data frame from \code{build_edges}}

\item{docs}{optional text data to include in results}
}
\description{
Similiar process to grouping vertices together that share edges into components
}
