% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measures.R
\name{jaccard}
\alias{jaccard}
\title{Fast jaccard similarity calculation.}
\usage{
jaccard(m)
}
\arguments{
\item{m}{- sparse \code{\link{dgCMatrix}} (or object that can be coerced to)
or dense input matrix.
\code{jaccard} calculates all pair similarities between rows of this matrix.}
}
\description{
Fast jaccard similarity calculation.
}
\examples{
i <- c(1,3:8, 1:6); j <- c(2,9,6:10,6:1); x <- c(6* (1:6), 7 * (1:7))
m <- sparseMatrix(i, j, x = x)
jaccard(m)
}

