% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measures.R
\name{jaccard_atomic}
\alias{jaccard_atomic}
\title{jaccard similarity between two vectors.}
\usage{
jaccard_atomic(x, y)
}
\arguments{
\item{x}{- \link{vector}}

\item{y}{- \link{vector}}
}
\value{
\link{numeric} jaccard similarity coefficient
}
\description{
jaccard similarity between two vectors.
}

