% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shingle.r
\name{jaccard_shingles}
\alias{jaccard_shingles}
\title{Jaccard similarity for shingled documents}
\usage{
jaccard_shingles(x, y)
}
\arguments{
\item{x, y}{shingle documents generated by \code{shingle_} function}
}
\description{
Jaccard similarity for shingled documents
}
