% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{jaccard_signatures}
\alias{jaccard_signatures}
\title{similarity between two signatures}
\usage{
jaccard_signatures(x, y)
}
\arguments{
\item{x}{- \link{vector}}

\item{y}{- \link{vector}}
}
\value{
\link{numeric} cosine similarity coefficient
}

