% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix.R
\name{lil_to_ngC}
\alias{lil_to_ngC}
\title{Convert binary Document-Term matrix from List-of-Lists
format to \code{\link{ngCMatrix-class}}
#'}
\usage{
lil_to_ngC(m_lil)
}
\description{
This function is specially useful when you want to do LSH for
"cosine" similarity (which current sketching algorithm work with ngCMatrix input)
}
\examples{
elems <- sapply(1:1000, function(z)
paste(sample(letters, sample(3:7), replace = T), collapse=''))
# generate artificial sets
sets <-  lapply(1:1e5, function(z) sample(elems, sample(20:100)))
system.time(dtm_lil <- create_dtm(sets, format = "lil"))
}

