% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsh.r
\name{lsh}
\alias{lsh}
\title{Locality Sensitivity Hashing}
\usage{
lsh(s, hashes = 100, rows = 5)
}
\arguments{
\item{s}{shingled document}

\item{hashes}{number of hash functions}

\item{rows}{number of rows per bands}
}
\description{
Locality Sensitivity Hashing
}
