% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lshr-package.R
\docType{package}
\name{lshr-package}
\alias{lshr}
\alias{lshr-package}
\title{lshr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Robert Zamora \email{zamora.rr@gmail.com}

}
\keyword{internal}
