% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhash.r
\name{minhash}
\alias{minhash}
\title{Minhash Document Shingles}
\usage{
minhash(shingles, n = 100, seed = NULL)
}
\arguments{
\item{shingles}{list of shingled documents (from \code{shingle_*()} functions)}

\item{n}{number of minhashes per document}

\item{seed}{optional seed used when generating minhash functions}
}
\description{
Minhash Document Shingles
}
