% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhash.r
\name{minhash_generator}
\alias{minhash_generator}
\title{Generate a minhash function}
\usage{
minhash_generator(n, seed = NULL)
}
\arguments{
\item{n}{number of minhashes per document}

\item{seed}{optional seed used when generating minhash functions}
}
\description{
Generate a minhash function
}
