% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix.R
\name{ngC_to_lil}
\alias{ngC_to_lil}
\title{Convert binary Document-Term matrix from \code{\link{ngCMatrix-class}}
format List-of-Lists}
\usage{
ngC_to_lil(dtm)
}
\description{
This function is specially useful when you want to do LSH for
"jaccard" similarity (which current minhashing algorithm work with List-of-Lists input)
}
\examples{
elems <- sapply(1:1000, function(z)
paste(sample(letters, sample(3:7), replace = T), collapse=''))
# generate artificial sets
sets <-  lapply(1:1e5, function(z) sample(elems, sample(20:100)))
system.time(dtm_ngC <- create_dtm(sets, format = "ngC"))
}

