% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.r
\name{prune_edges}
\alias{prune_edges}
\title{Prune results}
\usage{
prune_edges(edges, shingles, threshold = 0.8)
}
\arguments{
\item{edges}{from \code{build_edges}}

\item{shingles}{from \code{shingle}}
}
\description{
Apply jaccard similarity to candidates
}
