% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhash.r
\name{random_ints}
\alias{random_ints}
\title{Generate n random integers}
\usage{
random_ints(n, seed = NULL)
}
\arguments{
\item{n}{number of integers}

\item{seed}{optional seed}
}
\description{
Used to parametrize the minhasher
}
