% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.r
\name{tidy_candidates}
\alias{tidy_candidates}
\title{Tidy candidate results}
\usage{
tidy_candidates(candidates, shingles, docs = NULL, threshold = 0.8)
}
\arguments{
\item{candidates}{list of buckets with document ids from \code{lsh}}

\item{shingles}{list of documents and their shingles from \code{shingle}}

\item{docs}{optional text data to include in results}

\item{threshold}{jaccard similarity threshold}
}
\description{
Uses \code{jaccard_shingles} to compare candidate pairs from \code{lsh} to determine
which documents are actually similar.
}
