% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhashing.R
\name{to_lil}
\alias{to_lil}
\title{Converts 'dgCMatrix' to 'lil' format}
\usage{
to_lil(dtm)
}
\arguments{
\item{dtm}{Document-Term matrix}
}
\description{
Converts 'dgCMatrix' (or coercible to 'dgCMatrix') to 'lil'
format. This function is specially useful when you want to do LSH for
"jaccard" similarity (which current minhashing algorithm work with
List-of-Lists input)
}

