% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_candidate_pairs}
\alias{validate_candidate_pairs}
\title{Validate candidate pairs - remove false positive candidates}
\usage{
validate_candidate_pairs(candidate_indices, signature_matrix, similarity)
}
\arguments{
\item{candidate_indices}{- \link{data.table} (or \link{data.frame}) containing columns
\code{index1}, \code{index2} - indices of documents in candidate pair}

\item{signature_matrix}{input signature matrix - \code{\link{integer}} \code{\link{matrix}}}

\item{similarity}{similarity threshold. Pairs with signatures similarity >= threshold
will become near-neighbors.
Pairs with signatures similarity < threshold will treated as false positives and filtered out.}
}
\value{
\link{data.table} (or \link{data.frame}) containing columns
\code{index1}, \code{index2} - indices of similar documents without false positives
}
\description{
Validate candidate pairs - remove false positive candidates
}

