% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{cosine_atomic}
\alias{cosine_atomic}
\title{cosine similarity between two vectors.}
\usage{
cosine_atomic(x, y)
}
\arguments{
\item{x}{- \link{vector}}

\item{y}{- \link{vector}}
}
\value{
\link{numeric} cosine similarity coefficient
}

