% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{get_hash_matrix}
\alias{get_hash_matrix}
\title{Calculating hashing matrix for given similarity measure.}
\usage{
get_hash_matrix(dtm, hashfun_number, measure = c("jaccard", "cosine"),
  seed = 0L)
}
\arguments{
\item{hashfun_number}{- number of projections to use for approximation. Involves
tradeoff: larger is better, but computattionally more expensive and requires more RAM}

\item{measure}{Similarity measure}
}
\description{
For "jaccard" we generate construct hashing matrix
which simulate to permutation of rows. For "cosine" we generate random
hyperplanes.
}

