% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signatures.R
\name{get_signature_matrix}
\alias{get_signature_matrix}
\title{Calculating signature matrix using minhashing techniqe.}
\usage{
get_signature_matrix(tdm_lil, hashfun_number, measure = "jaccard",
  cores = parallel::detectCores())
}
\arguments{
\item{tdm_lil}{- sparse input matrix in list of lists form: \code{\link{list}} of \code{\link{vector}}s
It also may contain ncol attribute - number of columns.}

\item{hashfun_number}{number of hash functions to calculate signature matrix.}

\item{measure}{measure we want to calculate: \code{\link{character}}. Possible values - \code{c('jaccard', 'cosine')}}

\item{cores}{number of CPU threads to use while calculating hashes}
}
\value{
The signature matrix - \code{\link{integer}} \code{\link{matrix}} with dimension \code{hashfun_number * length(tdm_lil)}
}
\description{
Calculating signature matrix using minhashing techniqe.
}
\examples{
sets <- lapply(1:10, function(x) sample(letters, sample(5:15)))
tdm_lil <- create_tdm(sets)
sm <- get_signature_matrix(tdm_lil, hashfun_number = 10, measure = 'jaccard', cores = 4)
}

