% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minhash.R
\name{get_signature_matrix}
\alias{get_signature_matrix}
\title{Calculating signature matrix using minhashing techniqe.}
\usage{
get_signature_matrix(sets, hashfun_number, cores = parallel::detectCores())
}
\arguments{
\item{sets}{input sets in a form of \code{\link{list}} of \code{\link{vector}}s}

\item{hashfun_number}{number of hash functions to calculate signature matrix.}

\item{cores}{number of CPU threads to use while calculating hashes}
}
\value{
The signature matrix - \code{\link{integer}} \code{\link{matrix}} with dimension \code{hashfun_number * length(sets)}
}
\description{
Calculating signature matrix using minhashing techniqe.
}
\examples{
sets <- lapply(1:10, function(x) sample(letters, sample(5:15)))
sm <- get_signature_matrix(sets, 10, cores = 4)
}

