% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsh.R
\name{get_similar_pairs}
\alias{get_similar_pairs}
\title{Calculating candidate pairs using locality sensitive hashing.}
\usage{
get_similar_pairs(signature_matrix, bands_number, similarity,
  validate = FALSE, verbose = TRUE)
}
\arguments{
\item{signature_matrix}{input signature matrix - \code{\link{integer}} \code{\link{matrix}}}

\item{bands_number}{number of bands for LSH algorithm.}

\item{similarity}{target value of similarity we are looking for.}

\item{validate}{whether to perform validation and \strong{filtering out false positives
based on signatires similarity}}

\item{verbose}{- \link{logical} print lsh process information.
(such as expected false positive rate, false negative rate,timings, etc.)}
}
\value{
pairs of candidates with similarity => \code{similarity} -
\code{\link{data.table}} with 3 colums: index1, index2, N -
index of first candidate, index of second candidate,
and number of buckets where they share same value. The latter provided only for information.
(Intutition is following: the bigger N - the stronger similarity)
}
\examples{
sets <- lapply(1:10, function(x) sample(letters, sample(5:15)))
# add set similar to first set to the end of list
sets <- c(sets, list(c(sets[[1]], sample(letters, 5))))
sm <- get_signature_matrix(sets, 12)
get_similar_pairs(sm, 6, 0.9)
}

