\name{lsh}
\alias{lsh}
\title{Construct a new LSH object}
\usage{
lsh(d, b, seed)
}
\arguments{
  \item{d}{The dimension of the input vectors.}

  \item{b}{The number of bits to use in the hash.}

  \item{seed}{The seed to use to generate the random
  transformation matrix T.}
}
\description{
Construct a new LSH object
}

