% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lshr-package.R
\docType{package}
\name{lshr-package}
\alias{lshr}
\alias{lshr-package}
\title{lshr: An attempt at LSH}
\description{
Attempt to do locality sensitive hashing
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zamorarr/lshr}
  \item Report bugs at \url{https://github.com/zamorarr/lshr/issues}
}

}
\author{
\strong{Maintainer}: Robert Zamora \email{zamora.rr@gmail.com}

}
\keyword{internal}
