% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_candidate_pairs}
\alias{validate_candidate_pairs}
\title{Validate candidate pairs - calculates actual distances on candidate pairs}
\usage{
validate_candidate_pairs(m, i1, i2, similarity_measure = c("cosine",
  "jaccard"))
}
\arguments{
\item{m}{- input matrix}

\item{i1}{first candidate indices}

\item{i2}{second candidate indices}
}

